refactor: Referee 클래스에서 View와 관련된 부분을 Game에서 하도록 수정, Game 클래스를 컨트롤러로 수정
